♻️ Refactor contact card dialog into modular components
