+++ app/components/card-list/contact-card-dialog.tsx
@@ -1,12 +1,5 @@
 import { useState, useEffect } from "react";
-import {
-  Edit,
-  Trash2,
-  AlertTriangle,
-  Plus,
-  X,
-  MessageCircle,
-} from "lucide-react";
+import { AlertTriangle } from "lucide-react";
 import {
   Dialog,
   DialogContent,
@@ -14,42 +7,18 @@ import {
   DialogTitle,
   DialogFooter,
 } from "~/components/ui/dialog";
-import { Button } from "~/components/ui/button";
-import { Avatar, AvatarFallback, AvatarImage } from "~/components/ui/avatar";
-import { RoomAvatar } from "~/components/ui/room-avatar";
 import { Alert, AlertDescription } from "~/components/ui/alert";
-import { Input } from "~/components/ui/input";
-import { Label } from "~/components/ui/label";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "~/components/ui/form";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-import {
-  updateContactCard,
-  deleteContactCard,
-} from "~/lib/contacts-server-api/contacts";
-import {
-  getPlatformContacts,
-  createPlatformContact,
-  deletePlatformContact,
-} from "~/lib/contacts-server-api/platform-contacts";
+import { deleteContactCard } from "~/lib/contacts-server-api/contacts";
 import type {
   ContactCard as ContactCardType,
   PlatformContact,
-  ContactCardUpdate,
 } from "~/lib/contacts-server-api/types";
 import { useRoomContext } from "~/contexts/room-context";
-import { getDMRooms, type DMRoomInfo } from "~/lib/dm-room-utils";
-import { DMRoomSelector } from "~/components/ui/dm-room-selector";
-import { BridgeIcon } from "~/components/ui/bridge-icon";
-import { avatarFallback } from "~/lib/utils";
+import { getDMRooms } from "~/lib/dm-room-utils";
+import { ContactBasicInfo } from "./contact-basic-info";
+import { PlatformContactsList } from "./platform-contacts-list";
+import { DeleteConfirmation } from "./delete-confirmation";
+import { usePlatformContacts } from "~/hooks/use-platform-contacts";
 
 interface ContactCardDialogProps {
   readonly contactCard: ContactCardType | null;
@@ -63,24 +32,6 @@ interface ContactCardDialogProps {
   ) => void;
 }
 
-const contactFormSchema = z.object({
-  contact_name: z
-    .string()
-    .min(1, "Display name is required")
-    .max(100, "Display name must be less than 100 characters"),
-  nickname: z
-    .string()
-    .max(50, "Nickname must be less than 50 characters")
-    .optional(),
-  contact_avatar_url: z.string().optional(),
-});
-
-const platformFormSchema = z.object({
-  platform: z.enum(["Discord", "Telegram", "Matrix"]),
-  platform_user_id: z.string().min(1, "Platform user ID is required"),
-  dm_room_id: z.string().min(1, "DM room ID is required"),
-});
-
 export default function ContactCardDialog({
   contactCard,
   open,
@@ -90,82 +41,29 @@ export default function ContactCardDialog({
   onPlatformContactsUpdated,
 }: ContactCardDialogProps) {
   const { rooms } = useRoomContext();
-  const [isEditing, setIsEditing] = useState(false);
-  const [platformContacts, setPlatformContacts] = useState<PlatformContact[]>(
-    [],
-  );
-  const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [isDeleting, setIsDeleting] = useState(false);
   const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-  const [showDeletePlatformConfirm, setShowDeletePlatformConfirm] = useState<
-    string | null
-  >(null);
-  const [selectedDMRoom, setSelectedDMRoom] = useState<DMRoomInfo | null>(null);
-  const [showAddPlatform, setShowAddPlatform] = useState(false);
 
   const dmRooms = getDMRooms(rooms);
 
-  const form = useForm<z.infer<typeof contactFormSchema>>({
-    resolver: zodResolver(contactFormSchema),
-    defaultValues: {
-      contact_name: "",
-      nickname: "",
-      contact_avatar_url: "",
-    },
+  const {
+    platformContacts,
+    isLoading,
+    addPlatformContact,
+    removePlatformContact,
+  } = usePlatformContacts({
+    contactCardId: contactCard?.id || "",
+    onPlatformContactsUpdated,
+    onError: setError,
   });
 
   useEffect(() => {
     if (contactCard && open) {
-      setIsEditing(false); // Reset edit state when opening dialog
-      setShowDeleteConfirm(false); // Reset delete confirm state when opening dialog
-      setShowDeletePlatformConfirm(null); // Reset delete platform confirm state when opening dialog
-      setShowAddPlatform(false); // Reset add platform form when opening dialog
-      setSelectedDMRoom(null); // Reset selected DM room when opening dialog
-      form.reset({
-        contact_name: contactCard.contact_name,
-        nickname: contactCard.nickname || "",
-        contact_avatar_url: contactCard.contact_avatar_url || "",
-      });
-      loadPlatformContacts();
+      setShowDeleteConfirm(false);
+      setError(null);
     }
-  }, [contactCard, open, form]);
-
-  const loadPlatformContacts = async () => {
-    if (!contactCard) return;
-
-    setIsLoading(true);
-    setError(null);
-    try {
-      const contacts = await getPlatformContacts(contactCard.id);
-      setPlatformContacts(contacts);
-      onPlatformContactsUpdated?.(contactCard.id, contacts);
-    } catch (err) {
-      console.error("Failed to load platform contacts:", err);
-      setError("Failed to load platform contacts");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const onSubmit = async (values: z.infer<typeof contactFormSchema>) => {
-    if (!contactCard) return;
-
-    setError(null);
-    try {
-      const updateData: ContactCardUpdate = {
-        contact_name: values.contact_name,
-        nickname: values.nickname || null,
-        contact_avatar_url: values.contact_avatar_url || null,
-      };
-      const updatedCard = await updateContactCard(contactCard.id, updateData);
-      onCardUpdated?.(updatedCard);
-      setIsEditing(false);
-    } catch (err) {
-      console.error("Failed to update contact card:", err);
-      setError("Failed to update contact card");
-    }
-  };
+  }, [contactCard, open]);
 
   const handleDelete = async () => {
     if (!contactCard) return;
@@ -184,88 +82,13 @@ export default function ContactCardDialog({
     }
   };
 
-  const handleAddPlatformContact = () => {
-    setShowAddPlatform(true);
-    setSelectedDMRoom(null);
-  };
-
-  const handleSaveNewPlatformContact = async () => {
-    if (!contactCard || !selectedDMRoom) return;
-
-    if (!selectedDMRoom.platformUserId) {
-      setError("Selected room does not have valid platform information");
-      return;
-    }
-
-    const newPlatformContact = {
-      platform: selectedDMRoom.platform,
-      platform_user_id: selectedDMRoom.platformUserId,
-      dm_room_id: selectedDMRoom.roomId,
-    };
-
-    const validation = platformFormSchema.safeParse(newPlatformContact);
-    if (!validation.success) {
-      setError("Invalid platform contact data");
-      return;
-    }
-
-    setError(null);
-    try {
-      const newContact = await createPlatformContact({
-        contact_card_id: contactCard.id,
-        ...newPlatformContact,
-      });
-      const updatedPlatformContacts = [...platformContacts, newContact];
-      setPlatformContacts(updatedPlatformContacts);
-      onPlatformContactsUpdated?.(contactCard.id, updatedPlatformContacts);
-      setShowAddPlatform(false);
-      setSelectedDMRoom(null);
-    } catch (err) {
-      console.error("Failed to create platform contact:", err);
-      setError("Failed to create platform contact");
-    }
-  };
-
-  const handleDeletePlatformContact = async (platformContactId: string) => {
-    if (!contactCard) return;
-
-    setError(null);
-    try {
-      await deletePlatformContact(platformContactId);
-      const updatedPlatformContacts = platformContacts.filter(
-        (contact) => contact.id !== platformContactId,
-      );
-      setPlatformContacts(updatedPlatformContacts);
-      onPlatformContactsUpdated?.(contactCard.id, updatedPlatformContacts);
-      setShowDeletePlatformConfirm(null);
-    } catch (err) {
-      console.error("Failed to delete platform contact:", err);
-      setError("Failed to delete platform contact");
-    }
-  };
-
   if (!contactCard) return null;
-
-  const displayName = contactCard.contact_name || "Contact Name";
 
   return (
     <Dialog open={open} onOpenChange={onOpenChange}>
       <DialogContent className="max-w-2xl max-h-[80vh] overflow-y-auto">
         <DialogHeader>
-          <DialogTitle className="flex items-center gap-4">
-            Contact Details
-            <div className="flex gap-2">
-              {!isEditing && (
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  onClick={() => setIsEditing(true)}
-                >
-                  <Edit className="size-4" />
-                </Button>
-              )}
-            </div>
-          </DialogTitle>
+          <DialogTitle>Contact Details</DialogTitle>
         </DialogHeader>
 
         {error && (
@@ -276,302 +99,29 @@ export default function ContactCardDialog({
         )}
 
         <div className="space-y-6">
-          {/* Avatar and Basic Info */}
-          <div className="flex items-center gap-6">
-            <Avatar className="size-24">
-              <AvatarImage
-                src={contactCard.contact_avatar_url || undefined}
-                alt={displayName}
-              />
-              <AvatarFallback className="text-3xl font-semibold">
-                {avatarFallback(displayName)}
-              </AvatarFallback>
-            </Avatar>
+          <ContactBasicInfo
+            contactCard={contactCard}
+            onCardUpdated={onCardUpdated}
+            onError={setError}
+          />
 
-            <div className="flex-1 space-y-4">
-              {isEditing ? (
-                <Form {...form}>
-                  <form
-                    onSubmit={form.handleSubmit(onSubmit)}
-                    className="space-y-4"
-                  >
-                    <FormField
-                      control={form.control}
-                      name="contact_avatar_url"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Avatar</FormLabel>
-                          <FormControl>
-                            <Input
-                              placeholder="Avatar URL (optional)"
-                              disabled
-                              {...field}
-                            />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name="contact_name"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>
-                            Display Name <span className="text-red-500">*</span>
-                          </FormLabel>
-                          <FormControl>
-                            <Input
-                              placeholder="Enter display name"
-                              maxLength={100}
-                              {...field}
-                            />
-                          </FormControl>
-                          <FormMessage />
-                          <div className="text-xs text-muted-foreground text-right mt-1">
-                            {field.value?.length || 0}/100
-                          </div>
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name="nickname"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Nickname</FormLabel>
-                          <FormControl>
-                            <Input
-                              placeholder="Enter nickname (optional)"
-                              maxLength={50}
-                              {...field}
-                            />
-                          </FormControl>
-                          <FormMessage />
-                          <div className="text-xs text-muted-foreground text-right mt-1">
-                            {field.value?.length || 0}/50
-                          </div>
-                        </FormItem>
-                      )}
-                    />
-
-                    <div className="flex gap-2">
-                      <Button
-                        type="submit"
-                        disabled={form.formState.isSubmitting}
-                      >
-                        {form.formState.isSubmitting
-                          ? "Saving..."
-                          : "Save Changes"}
-                      </Button>
-                      <Button
-                        type="button"
-                        variant="outline"
-                        onClick={() => setIsEditing(false)}
-                      >
-                        Cancel
-                      </Button>
-                    </div>
-                  </form>
-                </Form>
-              ) : (
-                <div className="space-y-2">
-                  <div>
-                    <h2 className="text-2xl font-semibold">{displayName}</h2>
-                    {contactCard.nickname && (
-                      <p className="text-lg text-muted-foreground">
-                        {contactCard.nickname}
-                      </p>
-                    )}
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* Platform Contacts */}
-          <div className="space-y-4">
-            <div className="flex items-center justify-between">
-              <h3 className="text-lg font-semibold">Platform Accounts</h3>
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={handleAddPlatformContact}
-              >
-                <Plus className="size-4 mr-2" />
-                Add Platform
-              </Button>
-            </div>
-
-            {isLoading ? (
-              <div className="flex items-center justify-center py-8">
-                <div className="animate-spin rounded-full h-6 w-6 border-2 border-primary border-r-transparent" />
-              </div>
-            ) : (
-              <div className="space-y-3">
-                {/* Add new platform contact form */}
-                {showAddPlatform && (
-                  <div className="p-3 border rounded-lg space-y-3">
-                    <div className="space-y-2">
-                      <Label>Select DM Room</Label>
-                      <DMRoomSelector
-                        dmRooms={dmRooms}
-                        value={selectedDMRoom?.roomId}
-                        onValueChange={setSelectedDMRoom}
-                        placeholder="Choose a DM room to add as platform contact..."
-                      />
-                    </div>
-
-                    {selectedDMRoom && (
-                      <div className="p-3 bg-muted rounded-lg">
-                        <div className="flex items-center gap-3 mb-2">
-                          <div className="relative">
-                            <RoomAvatar
-                              roomId={selectedDMRoom.roomId}
-                              roomName={selectedDMRoom.roomName}
-                              fallbackAvatarUrl={selectedDMRoom.roomAvatar}
-                              className="size-8"
-                            />
-                            <span className="absolute -bottom-1 -right-1 flex items-center justify-center w-4 h-4 bg-gray-800 rounded-full ring-1 ring-gray-900">
-                              <BridgeIcon
-                                platform={selectedDMRoom.platform}
-                                className="size-3 text-white"
-                                showMatrix={true}
-                              />
-                            </span>
-                          </div>
-                          <div>
-                            <div className="font-medium">
-                              {selectedDMRoom.roomName}
-                            </div>
-                            <div className="text-sm text-muted-foreground">
-                              {selectedDMRoom.platform} •{" "}
-                              {selectedDMRoom.platformUserId}
-                            </div>
-                          </div>
-                        </div>
-                        <div className="text-xs text-muted-foreground">
-                          Room ID: {selectedDMRoom.roomId}
-                        </div>
-                      </div>
-                    )}
-
-                    <div className="flex gap-2">
-                      <Button
-                        size="sm"
-                        onClick={handleSaveNewPlatformContact}
-                        disabled={!selectedDMRoom}
-                      >
-                        Add Platform
-                      </Button>
-                      <Button
-                        size="sm"
-                        variant="outline"
-                        onClick={() => {
-                          setShowAddPlatform(false);
-                          setSelectedDMRoom(null);
-                        }}
-                      >
-                        <X className="size-4" />
-                      </Button>
-                    </div>
-                  </div>
-                )}
-
-                {platformContacts.map((contact) => (
-                  <div
-                    key={contact.id}
-                    className="flex items-center justify-between p-3 border rounded-lg"
-                  >
-                    <div className="flex items-center gap-3">
-                      <div className="flex items-center justify-center size-8 bg-gray-800 rounded-full">
-                        <BridgeIcon
-                          platform={contact.platform}
-                          showMatrix={true}
-                        />
-                      </div>
-                      <div>
-                        <p className="font-medium">{contact.platform}</p>
-                        <p className="text-sm text-muted-foreground">
-                          {contact.platform_user_id}
-                        </p>
-                      </div>
-                    </div>
-                    {showDeletePlatformConfirm !== contact.id ? (
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => setShowDeletePlatformConfirm(contact.id)}
-                        className="text-destructive hover:text-destructive"
-                      >
-                        <Trash2 className="size-4" />
-                      </Button>
-                    ) : (
-                      <div className="flex gap-1">
-                        <Button
-                          variant="destructive"
-                          size="sm"
-                          onClick={() =>
-                            handleDeletePlatformContact(contact.id)
-                          }
-                        >
-                          Confirm
-                        </Button>
-                        <Button
-                          variant="outline"
-                          size="sm"
-                          onClick={() => setShowDeletePlatformConfirm(null)}
-                        >
-                          Cancel
-                        </Button>
-                      </div>
-                    )}
-                  </div>
-                ))}
-
-                {platformContacts.length === 0 && !showAddPlatform && (
-                  <div className="text-center py-8 text-muted-foreground">
-                    <MessageCircle className="size-12 mx-auto mb-3 opacity-50" />
-                    <p>No platform accounts added yet</p>
-                    <p className="text-sm">
-                      Click "Add Platform" to connect accounts
-                    </p>
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
+          <PlatformContactsList
+            platformContacts={platformContacts}
+            dmRooms={dmRooms}
+            isLoading={isLoading}
+            onAddPlatformContact={addPlatformContact}
+            onDeletePlatformContact={removePlatformContact}
+          />
         </div>
 
         <DialogFooter className="flex-col sm:flex-row gap-2">
-          {!showDeleteConfirm ? (
-            <Button
-              variant="destructive"
-              onClick={() => setShowDeleteConfirm(true)}
-              className="sm:mr-auto"
-            >
-              <Trash2 className="size-4 mr-2" />
-              Delete Card
-            </Button>
-          ) : (
-            <div className="flex gap-2 sm:mr-auto">
-              <Button
-                variant="destructive"
-                onClick={handleDelete}
-                disabled={isDeleting}
-              >
-                {isDeleting ? "Deleting..." : "Confirm Delete"}
-              </Button>
-              <Button
-                variant="outline"
-                onClick={() => setShowDeleteConfirm(false)}
-              >
-                Cancel
-              </Button>
-            </div>
-          )}
+          <DeleteConfirmation
+            showConfirm={showDeleteConfirm}
+            isDeleting={isDeleting}
+            onShowConfirm={() => setShowDeleteConfirm(true)}
+            onConfirmDelete={handleDelete}
+            onCancel={() => setShowDeleteConfirm(false)}
+          />
         </DialogFooter>
       </DialogContent>
     </Dialog>
